pyproject.toml
[tool.poetry]
name = "pysubs"
version = "0.1.5"
description = "Python application to create subtitles for uploaded video file."
authors = ["Plato Puthur <plato@techtuft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
setuptools-rust = "^1.5.2"
uvicorn = "^0.20.0"
python-dotenv = "^0.21.0"
boto3 = "^1.26.49"
pytest = "^7.2.0"
pytube = "^12.1.2"
tokenizers = "^0.13.2"
numpy = "1.24.1"
tqdm = "^4.64.1"
more-itertools = "^9.0.0"
transformers = "^4.25.1"
ffmpeg-python = "0.2.0"
pytest-cov = "^4.0.0"
pydantic = "^1.10.4"
firebase-admin = "^6.0.1"
openai-whisper = "^20230117"
starlette-validation-uploadfile = "^0.1.1"

[tool.poetry.dev-dependencies]
torch = "1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
